Move Entity class to contracts.
